+++ CHANGELOG.md
@@ -12,6 +12,9 @@ All notable, unreleased changes to this project will be documented in this file.
 ### Other changes
 
 - Reference attribute linking to product variants - #10468 by @IKarbowiak
+### GraphQL API
+
+- Add `taxExemptionManage` mutation - #10344 by @SzymJ
 
 # 3.7.0
 
@@ -44,7 +47,6 @@ All notable, unreleased changes to this project will be documented in this file.
 
 - Add support for `bcrypt` password hashes - #10346 by @pkucmus
 - Add the ability to set taxes configuration per channel in the Avatax plugin - #10445 by @mociepka
-
 
 # 3.6.0